+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### 🚀 Added
+- **Slash Command System**: Introduce slash command system registered to desktop app and telegram.
+
+### ⚡ Changed
+- **Context Compact System**: Optimize the context compressor. It can be triggered manually or when the estimated tokens beyond a threshold.
+
+
 ## [v0.4.0]
 
 ### 🚀 Added